rearrange models into mcp
