+++ crates/mcp-core/src/role.rs
