+++ crates/mcp-core/src/tool.rs
