--- crates/mcp-macros/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "mcp-macros"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-syn = { version = "2.0", features = ["full", "extra-traits"] }
-quote = "1.0"
-proc-macro2 = "1.0"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-mcp-core = { path = "../mcp-core" }
-async-trait = "0.1"
-schemars = "0.8"
-
-[dev-dependencies]
-tokio = { version = "1.0", features = ["full"] }
-async-trait = "0.1"
-serde_json = "1.0"
-schemars = "0.8"
